javascript added
The code runs using a mouseOver event on the HTML id Element.

You can set the interval time frame to pick how fast the bar goes up.

We have the frame function clearing the interval every time it reaches
100.

Although, I have an issue where it begins to add another bar as one is
progressing, if you go away and return the mouse